Complete Epic: Semi-Autonomous Hill-Climbing Iteration System

This epic implements a full feedback loop for automated pattern algorithm
iteration using git, GitHub Actions, sensitivity analysis, and Pentagon
metrics evaluation.

Phases Completed:
- Phase 1: Foundation + Baseline (Tasks 56, 59, 61a, 61b)
- Phase 2: Iteration Core (Tasks 63, 66, 55, 64)
- Phase 3: Quality Gates + Expansion (Tasks 54, 61, 62)
- Phase 4: Visibility (Tasks 57, 58)
- Phase 5: Grid Expansion (Task 53)

Key Features:
- Weight-based algorithm blending with JSON configuration
- Sensitivity analysis identifies high-impact parameters
- /iterate command orchestrates single-pass improvements
- Ensemble search with tournament selection
- PR-based approval workflow with metric comparison
- 64-step pattern grid (expanded from 32)

All tests passing (373 test cases, 62,942 assertions).
Firmware: 92.12% flash, 5.81% SRAM.

Task 65 (Phrase-Aware Weights) deferred to backlog.

+++ src/Engine/EuclideanGen.h
@@ -40,7 +40,7 @@ namespace daisysp_idm_grids
  * - GenerateEuclidean(3, 8)  = 0x25       (son clave: 100100100)
  * - GenerateEuclidean(5, 8)  = 0xAA       (10101010)
  */
-uint32_t GenerateEuclidean(int hits, int steps);
+uint64_t GenerateEuclidean(int hits, int steps);
 
 /**
  * Rotate a pattern by offset steps (deterministic rotation).
@@ -53,7 +53,7 @@ uint32_t GenerateEuclidean(int hits, int steps);
  * Example:
  * - RotatePattern(0x11, 1, 8) = 0x88  (00010001 -> 10001000)
  */
-uint32_t RotatePattern(uint32_t pattern, int offset, int steps);
+uint64_t RotatePattern(uint64_t pattern, int offset, int steps);
 
 // =============================================================================
 // Euclidean + Probabilistic Weight Blending
@@ -81,11 +81,11 @@ uint32_t RotatePattern(uint32_t pattern, int offset, int steps);
  * @param seed RNG seed for Gumbel noise and rotation
  * @return Bitmask of selected hit steps
  */
-uint32_t BlendEuclideanWithWeights(
+uint64_t BlendEuclideanWithWeights(
     int budget,
     int steps,
     const float* weights,
-    uint32_t eligibility,
+    uint64_t eligibility,
     float euclideanRatio,
     uint32_t seed);
